Implemented the inventory for the player/s pulling from the database.
Pulls the item name and image as well - We still need to load the image via asset loader.

Updated routing file on the backend to start having a player routing section.This will be utilized like an api to retrieve any quick data we need for the player.

Updated the status on the authentication responses. Realised I didn't need to put status:400 within the return object.

Put the latest SQL structure into the sql folder too.