Change the code for when a player clicks on an item within the inventory. It also is able to play item sounds as well if they exist.
Update the load sounds function to make sure it doesn't try to load already loaded sounds.

Update the inventory drawing function to also load inventory item sounds.

Updated the map rendering class so that it loads the map background sound if assigned on the map.

Remove dead player inventory class on front end.

Renamed some assets to be lowercase.

Updated the game socket so that it fires the inventoryUsed event as opposed to increaseHealth.